projects/requirements.txt
langchain
langchain-openai
python-dotenv
faiss-cpu
bs4
langchain-community
langchain-groq
fastapi
uvicorn
langserve[all]
transformers
langchain_huggingface